Remove old code, replace with URL to APCA repo
It came to my attention that people are still using this page via GitHub.io, and as all of this information was deleted from the FPWD, it is important that at least this page be current. This commit deletes the old obsolete sampleJS code, corrects some important values in the plain language tutorial, and provides the link to the GitHub repo for the authoritative APCA code and methods.